feat: floxEnv: enforce unique packages
Don't allow multiple packages to install the same set of store paths

Note that partial overlap is still allowed; e.g. [out man] and [out lib]
is permitted even if both out paths are identical

If a user installs a package from unstable but then they want to switch
to stable, there should be a collision even if the storePaths are
identical. Otherwise a collision may occur later on during a flox
upgrade, and it would be confusing that both stable and unstable were
installed

The only time I can think of it being desirable to have two different
references to the same package would be:

floxEnv 1:
nixpkgs-flox.hello = {};

floxEnv 2:
curated-nixpkgs-flox.hello = {};

These two floxEnvs could not be composed, but this seems like a much
less common case than the undesirable case described above