fix: don't realize fakeDerivation for catalog.json
Currently determining the contents of manifest.json and catalog.json
requires realizing the storePath for a fakeDerivation. This is
undesirable since we should be able to eval a catalog without performing
any intensive operations in order to determine the available upgrades
for an environment. Instead of realizing the fakeDerivation storePath,
just pass through the storePaths that it wraps

Helps https://github.com/flox/product/issues/155

This means manifest.json and catalog.json may contain a storePaths with multiple storePaths rather than just a single one, and the output of flox list changes from:
Packages
    0  stable.nixpkgs-flox.curl  /nix/store/07zipzdqcldqz5f6wdarc2dlxmwbydi9-curl-7.86.0

To:
Packages
    0  stable.nixpkgs-flox.curl  /nix/store/8nv1g4ymxi2f96pbl1jy9h625v2risd8-curl-7.86.0-bin,/nix/store/i6jgil3y6pdbgi29vrsz8frg0y4d02lp-curl-7.86.0-man