feat: link propagated build inputs in floxEnvs
This patch updates the floxEnv builder to add all packages installed
to a flox environment to the nix-support/propagated-user-env-packages
file and adds a customized version of the nixpkgs buildEnv builder
which recursively treats propagated-{build-inputs,user-env-packages}
the same.

The effect of this change is to render symbolic links for all files
both in directly installed packages _and_ their (recursive) propagated
build input packages, so that files from these inputs can be found in
their customary FHS location under $FLOX_ENV, and correspondingly means
that $FLOX_ENV can be used in other variables like $PYTHONPATH.

Notably, this does _not_ increase the number of packages in the closure,
it only increases the number of symbolic links found in the floxEnv
package itself.