fix: don't always pass "out" output to buildEnv
The jq package is unusual in that it has a completely empty "out"
output which is not rendered by default, and passing a nonexistent
path to buildenv.pl causes it to fail.

This patch updates the logic to instead iterate over the outputs
specified by meta.outputsToInstall rather than blindly passing the
"out" output to the buildEnv function.

N.B. I'm intentionally leaving the builtins.trace call commented
out in place to make it easier to debug in future.