Drop tracelinks
Because this package is defined in a private repository which requires an authorized SSH token, unauthorized users encounter errors attempting to search in `floxpkgs`.
`pkgdb` simply ignores the error for this package and continues; but attempts to use things like `flake archive` or `eval` on this flake are not possible in earlier releases of `nix` which does not fetch flakes lazily.

`tracelinks` was previously used by `flox-bash` but has been deprecated. It is no longer in use by any `flox` projects that I am aware of.